refactor: split the PDS3 dataset tests by subject
The module reached 1109 lines, having gained the index-backed enumeration
tests this phase adds.  What a label and an index row say about one image
-- which file the label really names, the epoch and the camera the row
carries -- is a subject of its own: it is answered without SPICE and
without opening the image, and it is not about which images a selection
yields.  It moves to a module of its own, leaving 835 lines behind.

Every test moves byte for byte: 47 before, 47 after, no name changed and
no body changed.  The dataset object both modules ask their questions of
is the one thing they share, so it moves to the directory's conftest.
